LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-event-after-navigation-within-page.https-expected.txt



FAIL Service Worker should respond to fetch event after the hash changes assert_equals: expected "intercepted by service worker" but got ""
FAIL Service Worker should respond to fetch event after the pushState assert_equals: expected "intercepted by service worker" but got ""
